Improved BGP session handling
This commit introduces a flexible PeerConfig, using the new BgpConfig
trait from rotonda-fsm. It enables the user to specify prefixes in
addition to exact IP addresses for the remote side of a BGP session.
Also, multiple ASNs or 'any ASN' can be configured.

For all sessions, a channel to send commands to that session is stored
in the rotonda unit, keyed on remote IP and ASN. This way we can
hopefully re-use this unit from the East side of a Rotonda setup, i.e.
for 'bgp-out'.

Known issues / limitations:
- We only listen for incoming connections;
- Metrics and status reporting are incomplete;
- Withdrawing all observed prefixes upon session teardown is still
  missing.