Initial MVP default config support.
When run without --config the embedded copy of etc/rotonda.conf is used instead, and targets.proxy is removed if --proxy-destination is not supplied on the command line.

If --bgp-listen or --bmp-listen are supplied on the command line they override the BMP and BGP listen settings in the embedded config.

- Allow BGP peers to be missing.