Prepare MQTT target for proper unit testing (#60)
This PR was intended to augment the unit tests of the MQTT target, in particular to test that the MQTT metrics work. However, the way in which the target uses the rumqttc library is not testable, it is tightly integrated into the target code and spread across multiple Tokio tasks. Additionally the current code is not capable of dynamic reconfiguration as units and targets in Rotonda should be. This PR therefore replaces the main `run()` fn with new code to resolve these issues, and also splits out the various different pieces of the code supporting the target into their own Rust modules. A subsequent PR will then actually add the missing unit tests, this PR lays the foundation for that work.

In particular this PR:
- Removes `src/common/target.rs` which provided a custom `From` impl that was used to make `Either` outputs from `select()` more readable using the Rotonda specific `TargetActivity` enum.
- Rewrites the `run()` and `message_publisher()` fns of the MQTT target to use the `tokio::select!` macro instead of the previous `select()` and `TargetActivity` and multiple Tokio task based approach, greatly simplifying the code.
- Further simplifies the code by moving rumqttc connection state handling into a new `Connection` type.
- Makes one minor improvement to MQTT metrics by making the connection established metric atually a 0/1 state flag rather than describing it wrongly as a counter (as it didn't count...).
-  Fixes an issue whereby during the wait period between reconnection attempts no Rotonda target commands were handled, preventing CTRL-C shutdown of the application for example.
- Introduces use of traits to enable replacement of key behaviour (such as actually calling out to the rumqttc event loop `poll()` fn) with alternate versions when under test.
- Requires MQTT client ID to meet rumqttc rules to prevent an rumqttc panic.
- Adds a first metrics test demonstrating the possibilities enabled by this PR.