Use OwnedPathAttributes from routecore
With that, we can get rid of the octseq dependency in rotonda itself
again. The Cargo.toml(s) of rotonda and mrtin now share some dependency
versions via workspace.dependencies (routecore and inetnum, for now).
The routecore version points to a git rev until a proper crate is
published (likely 0.5.0-rc2).